requirements.txt
streamlit
easyocr
transformers
torch
Pillow
